Rename CPP and HPP files
