+++ AirspaceConverterQt/AboutDialog.hpp
