+++ src/Airspace.hpp
@@ -13,7 +13,7 @@
 #pragma once
 #include <string>
 #include <vector>
-#include "Geometry.h"
+#include "Geometry.hpp"
 
 class Altitude {
 public:
