+++ src/RasterMap.hpp
